==> MEDICAL AI ANALYSE/1 - COLLECTE DE DE DONNEES (  PARTIE 1 , WEB SCRAPING ).ipynb <==
# <center>COLLECTE DE DONNEES, PARTIE 1: WEB SCRAPING.</center>

   ### <center>LES MALADIES SUIVANTES ONT ETE SCRAPEES SUR WIKIPEDIA.</center>
                                 NB: ces donnees representent seulement une partie 
                                        des maladies de notre dataset final

### <center>OBJECTIF: Receuilir le maximum de noms de maladies sur internet.</center>
___

### Importations

import requests
from bs4 import BeautifulSoup
import pandas as pd

### Receuillons la liste de noms de cancer.

url = "https://fr.wikipedia.org/wiki/Liste_de_cancers"
pages = requests.get(url)

if (pages.status_code ==200):
    parsePage = BeautifulSoup(pages.content,"lxml")
    p = parsePage.find("div",{"class":"mw-content-ltr mw-parser-output"})
    a = p.find_all("li")
    cancers = ["".join(t.text.split("(")[::2]) for t in a]
    print(cancers[:5])
else:
    print("ERREUR: Veuillez verifier l'url de connection ")

### Receuillons la de noms des maladies infectueuses.

url = "https://fr.wikipedia.org/wiki/Liste_des_maladies_infectieuses"
pages = requests.get(url)

if (pages.status_code ==200):
    parsePage = BeautifulSoup(pages.content,"lxml")
    elemet =parsePage.find_all("td")
    maladies_inf = ["".join(t.text.split("(")[::2]).replace(")", '').replace("\n", '') for t in elemet][::2]
    print(maladies_inf[:5])
else:
    print("ERREUR: Veuillez verifier l'url de connection ")

### Creation du dataset  

data2 = {
    'Cancers' : cancers
}

data1 = {
    "Maladies infectueuses" : maladies_inf
}

df2 = pd.DataFrame(data2)
df1 = pd.DataFrame(data1)

df = pd.concat([df1, df2], axis=1)

df.head()

### Exportation

# Exportation du DataFrame en fichier CSV

df.to_excel('Ressources/RessourcesCollecte1/maladies_infectieuses.xlsx', index=False)

# Exportation en fichier excel

df.to_csv('Ressources/RessourcesCollecte1/maladies_infectieuses.csv', index=False)

print("DataFrame exporté avec succès.")
