==> MEDICAL AI ANALYSE/2 - COLLECTE DE DONNEES (PARTIE2, CREATION DU DATASET FINAL).ipynb <==
# <center>COLLECTE DE DONNEES, PARTIE2: CREATION DU DATASET FINAL.</center>
                                  NB: ces donnees ont été receuillies grace a CHATGPT  

### <center>OBJECTIF: Creer un dataset avec toutes les informations receuillies.</center>
---

### Importations

import pandas as pd

### Creation des dictionnaires de donnees

data_d = [
    {
        'Maladie': 'Endométriose',
        'Symptomes': ['Douleurs pelviennes intenses',
                        'Douleurs pendant les règles',
                        'Douleurs pendant les rapports sexuels',
                        'Douleurs lombaires',
                        'Douleurs abdominales',
                        'Fatigue chronique',
                        'Saignements menstruels abondants',
                        'Saignements menstruels irréguliers',
                        'Problèmes de fertilité',
                        'Troubles gastro-intestinaux pendant les règles',
                        'Douleurs pendant la défécation',
                        'Douleurs pendant la miction',
                        "Douleurs pendant l'ovulation",
                        'Nausées pendant les règles',
                        'vomissements pendant les règles',
                        "Douleurs pendant l'exercice physique",
                        'Problèmes urinaires pendant les règles'],
        'Traitement': ['médicaments anti-inflammatoires', 'chirurgie', 'traitement hormonal'],
        'Examens': ['imagerie médicale (échographie pelvienne)', 'laparoscopie']
    },
    {
        'Maladie': 'Anémie ferriprive',
        'Symptomes': ['Fatigue intense',
                    'Faiblesse générale',
                    'Palpitations cardiaques',
                    'Essoufflement',
                    'Peau pâle',
                    'Maux de tête fréquents',
                    'Étourdissements',
                    'Irritabilité',
                    'Difficultés de concentration',
                    'Ongles cassants',
                    'Cheveux fragiles',
                    'Sensibilité au froid',
                    'Mains et pieds froids',
                    'Langue gonflée ou lisse',
                    'Pica'],
        'Traitement': ['suppléments de fer', 'ajustement du régime alimentaire', 'traitement de la cause sous-jacente'],
        'Examens': ['analyse de sang (taux de fer)', 'numération sanguine complète', 'ferritine']
    },
    {
        'Maladie': 'Trouble bipolaire',
        'Symptomes': ["Variations extrêmes d'humeur",
                        'Irritabilité ',
                        'insomnie',
                        'Parler rapidement et de manière ininterrompue',
                        'Pensées grandioses ou idées délirantes',
                        'idées délirantes',
                        'Augmentation des activités sexuelles',
                        'Prise de décisions impulsives',
                        'Périodes de dépression intense',
                        "Perte d'énergie",
                        'Sentiments de désespoir',
                        "Changements d'appétit",
                        'Changements de poids'
                        'Problèmes de concentration',
                        'Pensées suicidaires'],
        'Traitement': ["médicaments stabilisateurs de l'humeur", 'thérapie cognitivo-comportementale', 'thérapie électroconvulsive'],
        'Examens': ['évaluation psychiatrique', 'entretiens cliniques']
    },
    {
        'Maladie': 'Maladie de Lyme',
        'Symptomes': ['Éruption cutanée en forme de cible (érythème migrant)',
                        'Fatigue',
                        'Fièvre et frissons',
                        'Maux de tête',
                        'Douleurs musculaires et articulaires',
                        'Gonflement des ganglions lymphatiques',
                        'Raideur du cou',
                        'Troubles cognitifs',
                        'Problèmes de sommeil',
                        'Irritabilité',
                        'Sensibilité à la lumière et au bruit',
                        'Paralysie faciale',
                        'Problèmes cardiaques'],
        'Traitement': ['antibiotiques', 'médicaments anti-inflammatoires', 'repos'],
        'Examens': ["test sanguin (détection d'anticorps)", 'analyse de liquide articulaire', 'imagerie médicale']
    },
    {
        'Maladie': 'Schizophrénie',
        'Symptomes': [ 'Hallucinations auditives',
                      'Hallucinations olfactives',
                      'Hallucinations visuelles',
                        'Idées délirantes',
                        'Troubles de la pensée',
                        'Discours désorganisé',
                        'Émoussement affectif',
                        'Retrait social',
                        'Isolement social',
                        'Perturbation de la mémoire',
                        'Difficultés de concentration',
                        'Paranoïa',
                        'Problèmes de sommeil'],
        'Traitement': ['antipsychotiques', 'thérapie psychosociale', 'hospitalisation en cas de crise'],
        'Examens': ['évaluation psychiatrique', 'entretiens cliniques', 'tests psychologiques']
    },
    {
        'Maladie': 'Cancer du poumon',
        'Symptomes': ['Toux persistante',
                        'Essoufflement',
                        'Douleur thoracique',
                        'Perte de poids inexpliquée',
                        'Fatigue',
                        'Infections respiratoires fréquentes',
                        'Hoquet persistant',
                        'Sang dans les expectorations',
                        'Changements de voix',
                        'Difficultés à avaler',
                        'Gonflement des ganglions lymphatiques'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['imagerie médicale (scanner, IRM)', 'biopsie', 'analyse de crachat']
    },
    {
        'Maladie': 'LGV (Lymphogranulomatose vénérienne)',
        'Symptomes': ['Ulcères génitaux ou anaux',
                        'Douleur en urinant',
                        'Écoulement génital ou anal',
                        'Gonflement des ganglions lymphatiques inguinaux',
                        'Douleur abdominale basse',
                        'Saignement rectal',
                        'Diarrhée',
                        'Fièvre',
                        'Perte de poids',
                        'Fatigue'],
        'Traitement': ['antibiotiques', 'éviction sexuelle pendant le traitement', 'suivi médical régulier'],
        'Examens': ['prélèvement de liquide des ulcères', 'tests sanguins', 'imagerie médicale']
    },
    {
        'Maladie': 'Ureaplasma génital',
        'Symptomes': ['Douleur ou brûlure pendant la miction',
                        "Écoulement anormal du vagin ou de l'urètre",
                        'Douleur pelvienne',
                        'Douleur pendant les rapports sexuels',
                        'Irritation génitale',
                        'Saignements inhabituels',
                        "Inflammation ou irritation du col de l'utérus",
                        'Douleur ou inflammation articulaire',
                        'Fièvre',
                        'Fatigue',
                        'Sensibilité abdominale'],
        'Traitement': ['antibiotiques', 'éviction sexuelle pendant le traitement', 'traitement du partenaire sexuel'],
        'Examens': ['prélèvement génital', 'tests sanguins']
    },
    {
        'Maladie': 'Pneumocystose',
        'Symptomes': ['Toux sèche persistante',
                        'Essoufflement',
                        'Fièvre',
                        'Perte de poids inexpliquée',
                        'Fatigue extrême',
                        'Douleurs thoraciques',
                        'Sueurs nocturnes',
                        'Difficultés respiratoires',
                        'Respiration rapide',
                        "Taux d'oxygène bas dans le sang"],
        'Traitement': ['antibiotiques', 'traitement antiparasitaire', 'oxygénothérapie'],
        'Examens': ['examen microscopique des sécrétions respiratoires', 'analyse sanguine', 'radiographie thoracique']
    },
    {
        'Maladie': 'Toxoplasmose',
        'Symptomes': ['Gonflement des ganglions lymphatiques',
                        'Fièvre',
                        'Maillons',
                        'Fatigue',
                        'Maux de tête',
                        'Douleurs musculaires',
                        'Sensation de malaise général',
                        'Inflammation de la gorge',
                        'Éruptions cutanées',
                        'Douleurs abdominales'],
        'Traitement': ['antibiotiques', 'repos', 'traitement symptomatique'],
        'Examens': ["test sanguin (détection d'anticorps)", 'biopsie en cas de symptômes graves']
    },
    {
        'Maladie': 'Cryptosporidiose',
        'Symptomes': ['Diarrhée aqueuse',
                        'Douleurs abdominales',
                        'Fièvre',
                        'Nausées',
                        'Vomissements',
                        'Perte de poids',
                        'Fatigue',
                        'Douleurs musculaires',
                        'Maillons',
                        'Déshydratation'],
        'Traitement': ['réhydratation', 'antiparasitaires', 'traitement symptomatique'],
        'Examens': ['analyse de selles', 'tests sanguins', 'endoscopie en cas de symptômes graves']
    },
    {
        'Maladie': 'Leishmaniose',
        'Symptomes': ['Lésions cutanées (ulcères) sur la peau',
                        'Fièvre',
                        'Fatigue',
                        'Perte de poids',
                        'Gonflement des ganglions lymphatiques',
                        'Anémie',
                        'Douleurs musculaires',
                        'Douleurs articulaires',
                        'Sensations de brûlure',
                        'Sensations de picotement'],
        'Traitement': ['antiparasitaires', 'antibiotiques', 'soins de la plaie'],
        'Examens': ["analyse de tissu prélevé de l'ulcère", 'tests sanguins', 'imagerie médicale']
    },
    {
        'Maladie': 'Lèpre',
        'Symptomes': [ 'Lésions cutanées hypopigmentées ou érythémateuses',
                        'Engourdissement',
                        'Faiblesse musculaire',
                        'Déformations des membres',
                        'Douleurs nerveuses',
                        'Perte de sourcils ou de cils',
                        'Perte de la capacité à sentir la chaleur ou le froid',
                         "perte de sensation dans les membres"],
        'Traitement': ['multithérapie antibactérienne', 'soins de la peau', 'réadaptation physique'],
        'Examens': ['biopsie de la peau', 'examen des nerfs', 'tests sanguins']
    },
    {
        'Maladie': 'Tuberculose',
        'Symptomes': ['Toux persistante',
                        'Expectorations de sang',
                        'Fatigue',
                        'Fièvre',
                        'Perte de poids inexpliquée',
                        'Sueurs nocturnes',
                        'Essoufflement',
                        'Douleurs thoraciques',
                        'Appétit réduit'],
        'Traitement': ['antibiotiques (multithérapie)', 'isolement', 'traitement directement observé (TDO)'],
        'Examens': ['radiographie thoracique', 'test cutané à la tuberculine', 'analyse de crachats']
    },
    {
        'Maladie': 'Sarcoïdose',
        'Symptomes': ['Gonflement des ganglions lymphatiques',
                        'Essoufflement',
                        'Toux persistante',
                        'Fatigue',
                        'Douleurs articulaires',
                        'Éruptions cutanées',
                        'Fèces sanglantes ou glaireuses',
                        'Troubles de la vision',
                        'Douleurs thoraciques'],
        'Traitement': ['corticoïdes', 'immunosuppresseurs', 'traitement symptomatique'],
        'Examens': ['radiographie thoracique', 'tomographie par émission de positrons (TEP)', 'biopsie']
    },
    {
        'Maladie': 'Amibiase',
        'Symptomes': ['Diarrhée',
                        'Douleurs abdominales',
                        'Crampes abdominales',
                        'Nausées',
                        'Vomissements',
                        'Fatigue',
                        'Perte de poids',
                        'sèlles sanglantes ou glaireuses'],
        'Traitement': ['antiparasitaires', 'antibiotiques', 'réhydratation'],
        'Examens': ['analyse de selles', 'imagerie médicale', 'tests sanguins']
    },
    {
        'Maladie': 'Giardiase',
        'Symptomes': ['Diarrhée',
                        'Flatulences',
                        'Douleurs abdominales',
                        'Nausées',
                         "Perte d'appétit",
                        'Fatigue',
                        'Perte de poids',
                        'Ballonnements'],
        'Traitement': ['antiparasitaires', 'réhydratation', 'traitement symptomatique'],
        'Examens': ['analyse de selles', 'tests sanguins', 'imagerie médicale']
    },
    {
        'Maladie': 'SIDA',
        'Symptomes': ['Fatigue', 'Fièvre', 'Gonflement des ganglions lymphatiques',
                        'Perte de poids inexpliquée', 'Sueurs nocturnes', 'Maux de tête',
                        'Douleurs musculaires', 'Maux de gorge', 'Infections opportunistes',
                        'Diarrhée persistante', 'Toux prolongée', 'Lésions cutanées',
                        'Méningite', 'Problèmes de mémoire et de concentration'],
        'Traitement': ['traitement antirétroviral'],
        'Examens': ['test de dépistage du VIH', 'compte de lymphocytes CD4']
    },
    {
        'Maladie': 'Syphilis',
        'Symptomes': [ 'Chancre indolore (ulcération)',
                        'Éruption cutanée sur la peau, y compris les paumes et les plantes des pieds',
                        'Ganglions lymphatiques enflés',
                        'Fatigue',
                        'Fièvre',
                        'Maux de tête',
                        'Douleurs musculaires',
                        'Perte de poids inexpliquée',
                        'Saisons de fièvre, sueurs nocturnes et fatigue (stades avancés)'],
        'Traitement': ['pénicilline', 'doxycycline'],
        'Examens': ['test de la syphilis', 'examen au microscope']
    },
    {
        'Maladie': 'Chlamydia',
        'Symptomes': ['Douleur ou brûlure pendant la miction',
                        'Écoulement anormal du pénis ou du vagin',
                        'Douleur abdominale basse',
                        'Douleurs pendant les rapports sexuels',
                        'Saignements entre les règles',
                        'Douleurs testiculaires',
                        'Inflammation cervicale'],
        'Traitement': ['azithromycine', 'doxycycline'],
        'Examens': ['test de dépistage de la chlamydia', 'culture cellulaire']
    },
    {
        'Maladie': 'Gonorrhée',
        'Symptomes': ['Écoulement épais et jaune du pénis ou du vagin',
                        'Douleur ou brûlure pendant la miction',
                        'Douleurs abdominales basses',
                        'Douleurs pendant les rapports sexuels',
                        'Saignements entre les règles',
                        'Gorge irritée'],
        'Traitement': ['céfixime', 'azithromycine'],
        'Examens': ['test de dépistage de la gonorrhée', 'culture cellulaire']
    },
    {
        'Maladie': 'Herpès génital',
        'Symptomes': ['lésions génitales', 'démangeaisons', 'douleur en urinant',"Vésicules douloureuses sur les organes génitaux ou autour de l'anus",
                        'Démangeaisons',
                        'Sensation de brûlure pendant la miction',
                        'Douleurs pendant les rapports sexuels',
                        'Gonflement des ganglions lymphatiques'],
        'Traitement': ['antiviraux', 'médicaments symptomatiques'],
        'Examens': ["test de l'herpès", 'examen clinique']
    },
    {
        'Maladie': 'Condylomes génitaux (HPV)',
        'Symptomes': ['verrues génitales', 'démangeaisons','Saignements pendant les rapports sexuels'],
        'Traitement': ['crèmes antivirales', 'ablation des verrues'],
        'Examens': ['test HPV', 'examen visuel']
    },
    {
        'Maladie': 'Hépatite B',
        'Symptomes': ['Fatigue', 'Jaunisse', 'Douleurs abdominales',"Perte d'appétit", 'Nausées et vomissements',
                        'Urine foncée', 'Fièvre', 'Selles décolorées','Douleurs articulaires', 'Irritation cutanée'],
        'Traitement': ['antiviraux', "vaccination (pour l'hépatite B)"],
        'Examens': ["test d'hépatite", 'échographie hépatique']
    },
    {
        'Maladie': 'Hépatite C',
        'Symptomes': ['Fatigue', 'Jaunisse', 'Douleurs abdominales',
                        "Perte d'appétit", 'Nausées et vomissements',
                        'Urine foncée', 'Fièvre', 'Selles décolorées',
                        'Douleurs musculaires et articulaires', 'Irritation cutanée'],
        'Traitement': ['antiviraux', "vaccination (pour l'hépatite B)"],
        'Examens': ["test d'hépatite", 'échographie hépatique']
    },
    {
        'Maladie': 'Trichomonase',
        'Symptomes': ['démangeaisons', 'écoulement génital', 'douleur en urinant','Écoulement vaginal vert ou jaune-vert',
                        'Démangeaisons ou irritation vaginal',
                        'Douleur pendant les rapports sexuels',
                        'Douleur pendant la miction',
                        'Gonflement des organes génitaux externes'],
        'Traitement': ['métronidazole', 'tinidazole'],
        'Examens': ['test de dépistage de la trichomonase', 'examen microscopique']
    },
    {
        'Maladie': 'Molluscum contagiosum',
        'Symptomes': ['lésions cutanées', 'démangeaisons','Petites bosses ou nodules sur la peau',
                        'Aspect perlé ou lisse des lésions',
                        'Inflammation des lésions'],
        'Traitement': ['ablation des lésions', 'médicaments antiviraux'],
        'Examens': ['examen visuel', 'biopsie (dans certains cas)']
    },
    {
        'Maladie': 'Mycoplasmose génitale',
        'Symptomes': ['Douleur ou brûlure pendant la miction',
                    'Écoulement génital anormal',
                    'Douleur pelvienne',
                    'Douleur pendant les rapports sexuels',
                    'Irritation génitale',
                    'Saignements entre les règles',
                    'Inflammation cervicale',
                    'Douleur abdominale basse',
                    'Fièvre',
                    'Fatigue'],
        'Traitement': ['azithromycine', 'doxycycline'],
        'Examens': ['test de dépistage de la mycoplasmose', 'culture cellulaire']
    },
    {
        'Maladie': 'Ureaplasmose',
        'Symptomes': ['Douleur ou brûlure pendant la miction',
                        "Écoulement anormal du vagin ou de l'urètre",
                        'Douleur pelvienne',
                        'Douleur pendant les rapports sexuels',
                        'Irritation génitale',
                        'Saignements inhabituels',
                        'Douleur ou inflammation articulaire',
                        'Fièvre',
                        'Fatigue',
                        'Sensibilité abdominale'],
        'Traitement': ['azithromycine', 'doxycycline'],
        'Examens': ["test de dépistage de l'ureaplasmose", 'culture cellulaire']
    },
    {
        'Maladie': 'Chancre mou (Ulcère molle)',
        'Symptomes': ['Ulcération génitale indolore',
                        'Ganglions lymphatiques enflés',
                        'Fièvre',
                        'Douleurs musculaires',
                        'Maux de tête',
                        'Fatigue',
                        'Éruptions cutanées'],
        'Traitement': ['azithromycine', 'céfixime'],
        'Examens': ['test de dépistage du chancre mou', 'examen microscopique']
    },
    {
        'Maladie': 'Granulome inguinal (Donovanose)',
        'Symptomes': [ 'Ulcération génitale indolore',
                        'Ganglions lymphatiques enflés',
                        'Douleur pendant les rapports sexuels',
                        'Saignements pendant les rapports sexuels',
                        'Éruptions cutanées',
                        'Nodules ou bosses génitales'],
        'Traitement': ['doxycycline', 'azithromycine'],
        'Examens': ['test de dépistage de la donovanose', 'examen microscopique']
    },
    {
        'Maladie': 'Lymphogranulomatose vénérienne (LGV)',
        'Symptomes': ['Petite lésion ou ulcération génitale',
                        "Ganglions lymphatiques enflés dans l'aine",
                        'Douleur pendant les rapports sexuels',
                        'Saignements pendant les rapports sexuels',
                        'Douleur abdominale basse',
                        'Fièvre',
                        'Fatigue'],
        'Traitement': ['doxycycline', 'azithromycine'],
        'Examens': ['test de dépistage de la LGV', 'culture cellulaire']
    },
    {
        'Maladie': 'Mycobacterium avium complex (MAC)',
        'Symptomes': ['Fièvre',
                        'Sueurs nocturnes',
                        'Fatigue',
                        'Perte de poids',
                        'Diarrhée persistante',
                        'Douleurs abdominales',
                        'Infections opportunistes'],
        'Traitement': ['antibiotiques', 'traitement symptomatique'],
        'Examens': ['test de dépistage du MAC', 'imagerie médicale']
    },
    {
        'Maladie': 'Infections à clamidies non génitales',
        'Symptomes': [ 'Infections oculaires (conjonctivite)',
                        'Infections respiratoires',
                        'Douleurs articulaires',
                        'Frissons',
                        'Fièvre',
                        'Maillons'],
        'Traitement': ['azithromycine', 'doxycycline'],
        'Examens': ['test de dépistage des infections à chlamydia', 'culture cellulaire']
    },
    {
        'Maladie': 'Tréponématose non vénérienne',
        'Symptomes': ['Lésions cutanées (chancre)',
                        'Ganglions lymphatiques enflés',
                        'Fièvre',
                        'Fatigue',
                        'Maux de tête',
                        'Douleurs musculaires',
                        'Éruptions cutanées'],
        'Traitement': ['pénicilline', 'doxycycline'],
        'Examens': ['test de dépistage de la tréponématose non vénérienne', 'examen microscopique']
    },
    {
        'Maladie': 'Diabète de type 2',
        'Symptomes': ['Polyurie (urination fréquente)',
                        'soif excessive',
                        'faim excessive',
                        'Fatigue',
                        'Vision floue',
                        'Cicatrisation lente des plaies',
                        'Engourdissement ou picotements dans les pieds ou les mains',
                        'Infections fréquentes'],
        'Traitement': ['médicaments hypoglycémiants', 'régime alimentaire', 'exercice physique'],
        'Examens': ['glycémie à jeun', 'hémoglobine A1c']
    },
    {
        'Maladie': 'Hypertension artérielle',
        'Symptomes': [ 'Maux de tête fréquents',
                        'Vertiges',
                        'Essoufflement',
                        'Fatigue',
                        'Vision floue',
                        'Saignements de nez fréquents',
                        'Battements cardiaques irréguliers',
                        'Étourdissements'],
        'Traitement': ['antihypertenseurs', 'changement de mode de vie'],
        'Examens': ['mesure de la pression artérielle', 'analyse d\'urine']
    },
    {
        'Maladie': 'Cancer du sein',
        'Symptomes': ["Nodule ou masse dans le sein ou l'aisselle",
                        'Changement de taille ou de forme du sein',
                        'Douleur au sein ou au mamelon',
                        'Écoulement du mamelon (autre que le lait maternel)',
                        'Peau du sein rouge, épaissie ou présentant des écailles',
                        "Changements dans l'apparence du mamelon ou de l'aréole"],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['mammographie', 'biopsie']
    },
    {
        'Maladie': 'Glaucome',
        'Symptomes': ['Vision floue ou réduite',
                        'Vision des halos autour des lumières',
                        'Maux de tête',
                        'Nausées ou vomissements',
                        'Œil rouge',
                        'Perte de vision latérale (vision périphérique)'],
        'Traitement': ['collyres', 'chirurgie au laser', 'chirurgie traditionnelle'],
        'Examens': ['mesure de la pression intraoculaire', 'examen du champ visuel', 'gonioscopie']
    },
    {
        'Maladie': 'Leucémie',
        'Symptomes': ['Fatigue intense',
                        'Perte de poids inexpliquée',
                        'Fièvre persistante ou fréquente',
                        'Infections fréquentes',
                        'Saignements ou ecchymoses faciles',
                        'Ganglions lymphatiques enflés',
                        'Douleurs osseuses ou articulaires',
                        'Anémie'],
        'Traitement': ['chimiothérapie', 'greffe de moelle osseuse', 'thérapie ciblée'],
        'Examens': ['analyse de sang (numération sanguine complète)', 'biopsie de la moelle osseuse', 'imagerie médicale']
    },
    {
        'Maladie': 'Maladie de Crohn',
        'Symptomes': ['Diarrhée fréquente',
                        'Douleurs abdominales',
                        'Perte de poids inexpliquée',
                        'Fatigue',
                        'Fièvre',
                        'Saignements rectaux',
                        'Inflammation des articulations',
                        'Ulcérations buccales'],
        'Traitement': ['anti-inflammatoires', 'immunosuppresseurs', 'chirurgie'],
        'Examens': ['endoscopie', 'coloscopie', 'imagerie par résonance magnétique (IRM)']
    },
    {
        'Maladie': "Maladie d'Alzheimer",
        'Symptomes': ['Perte de mémoire',
                        'Difficulté à se rappeler des informations récentes',
                        'Confusion mentale',
                        'Difficulté à effectuer des tâches familières',
                        'Changements de personnalité',
                        'Difficulté à trouver des mots',
                        'Problèmes de jugement'],
        'Traitement': ['médicaments antidémence', 'thérapie occupationnelle', 'soutien psychologique'],
        'Examens': ['évaluation cognitive', 'imagerie cérébrale', 'analyses de sang']
    },
    {
        'Maladie': 'Maladie de Parkinson',
        'Symptomes': ['Tremblements au repos',
                        'Raideur musculaire',
                        'Lenteur des mouvements',
                        'Difficulté à commencer ou à arrêter un mouvement',
                        "Problèmes d'équilibre",
                        "Changements de l'expression faciale (masque parkinsonien)"],
        'Traitement': ['médicaments dopaminergiques', 'thérapie physique', 'chirurgie cérébrale profonde'],
        'Examens': ['examen neurologique', 'imagerie cérébrale', 'tests sanguins']
    },
    {
        'Maladie': 'Asthme',
        'Symptomes': ['Essoufflement',
                        'Toux fréquente, en particulier la nuit',
                        'Sifflements ou respiration sifflante',
                        'Sensation de serrement dans la poitrine',
                        'Fatigue due à un effort physique'],
        'Traitement': ['inhalateurs de corticostéroïdes', 'bronchodilatateurs', 'thérapie respiratoire'],
        'Examens': ['spirométrie', "tests d'allergie", 'radiographie thoracique']
    },
    {
        'Maladie': 'Maladie cœliaque',
        'Symptomes': ['Diarrhée chronique',
                        'Douleurs abdominales',
                        'Ballonnements',
                        'Fatigue',
                        'Perte de poids inexpliquée',
                        'Anémie',
                        'Irritabilité'],
        'Traitement': ['régime sans gluten', 'suppléments nutritionnels', 'médicaments symptomatiques'],
        'Examens': ['analyse de sang (anticorps)', "biopsie de l'intestin grêle", 'tests génétiques']
    },
    {
        'Maladie': 'Fibromyalgie',
        'Symptomes': ['Douleurs musculaires diffuses',
                        'Fatigue généralisée',
                        'Troubles du sommeil',
                        'Raideur matinale',
                        'Problèmes de concentration (brouillard cérébral)',
                        'Maux de tête'],
        'Traitement': ['médicaments analgésiques', 'thérapie physique', 'thérapie cognitivo-comportementale'],
        'Examens': ['aucun test spécifique', "diagnostic basé sur les symptômes"]
    },
    {
        'Maladie': 'Arthrite rhumatoïde',
        'Symptomes': ['Douleurs articulaires',
                        'Raideur articulaire',
                        'Enflure des articulations',
                        'Fatigue',
                        'Fièvre',
                        "Perte d'appétit",
                        'Nodules sous la peau'],
        'Traitement': ['médicaments anti-inflammatoires', 'thérapie physique', 'médicaments antirhumatismaux'],
        'Examens': ['analyse de sang (facteur rhumatoïde)', 'radiographie articulaire', 'échographie articulaire']
    },
    {
        'Maladie': 'Maladie cardiaque coronarienne',
        'Symptomes': ['Douleur thoracique ou inconfort (angine de poitrine)',
                        'Essoufflement',
                        'Fatigue',
                        'Nausées',
                        'Sueur excessive',
                        'Douleur ou engourdissement dans les bras, le cou, la mâchoire ou le dos'],
        'Traitement': ['médicaments anti-angineux', 'angioplastie coronarienne', 'bypass coronarien'],
        'Examens': ['électrocardiogramme (ECG)', 'échocardiographie', 'cathétérisme cardiaque']
    },
    {
        'Maladie': 'Sclérose en plaques',
        'Symptomes': ['Fatigue',
                        'Problèmes de vision',
                        'Engourdissements ou picotements',
                        'Faiblesse musculaire',
                        'Difficulté à coordonner les mouvements',
                        'Troubles de la mémoire et de la concentration'],
        'Traitement': ['médicaments immunomodulateurs', 'thérapie physique', 'médicaments symptomatiques'],
        'Examens': ['IRM cérébrale et médullaire', 'ponction lombaire', 'électromyographie (EMG)']
    },
     {
        'Maladie': 'Tumeur du cerveau',
        'Symptomes': ['Maux de tête persistants',
                        'Nausées et vomissements',
                        'Convulsions',
                        'Changements dans la vision',
                        "Problèmes d'équilibre et de coordination",
                        "Changements de personnalité ou d'humeur",
                        'Confusion ou désorientation',
                        'Problèmes de mémoire',
                        'Engourdissement ou faiblesse dans les bras ou les jambes',
                        'Troubles du sommeil',
                        "Difficultés d'élocution",
                        'Vision floue ou double',
                        "Changements dans la sensation ou le mouvement d'une partie du corps"],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['imagerie cérébrale (IRM et scanner)', 'biopsie', 'analyse sanguine']
    },
    {
        'Maladie': 'Cancer du col utérin',
        'Symptomes': ['Saignements vaginaux anormaux',
                        'Douleur pendant les rapports sexuels',
                        'Douleur pelvienne',
                        'Fatigue',
                        'Perte de poids inexpliquée',
                        'Écoulement vaginal anormal'],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['frottis cervico-utérin', 'colposcopie', 'biopsie']
    },
    {
        'Maladie': 'Cancer colorectal',
        'Symptomes': ['Changements dans les habitudes intestinales',
                        'Sang dans les selles',
                        'Douleur abdominale',
                        'Fatigue',
                        'Perte de poids inexpliquée',
                        'Anémie'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['coloscopie', 'sigmoïdoscopie', 'analyses de sang']
    },
    {
        'Maladie': "Cancer de l'endomètre",
        'Symptomes': ['Saignements vaginaux anormaux',
                        'Douleur pelvienne',
                        'Douleur pendant les rapports sexuels',
                        'Fatigue',
                        'Perte de poids inexpliquée'],
        'Traitement': ['chirurgie', 'radiothérapie', 'hormonothérapie'],
        'Examens': ["biopsie de l'endomètre", 'échographie pelvienne', 'imagerie médicale']
    },
                    
     {
        'Maladie': "Cancer de l'estomac",
        'Symptomes': ['Douleur abdominale',
                            'Sensation de plénitude après un repas',
                            "Perte d'appétit",
                            'Perte de poids inexpliquée',
                            'Difficulté à avaler'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['endoscopie gastrique', 'biopsie', 'imagerie médicale']
    },
    {
        'Maladie': 'Cancer du foie',
        'Symptomes': ['Douleur abdominale',
                        "Gonflement de l'abdomen",
                        'Jaunissement de la peau et des yeux (ictère)',
                        'Perte de poids inexpliquée',
                        'Fatigue'],
        'Traitement': ['chirurgie', 'transplantation hépatique', 'chimioembolisation'],
        'Examens': ['échographie abdominale', 'IRM du foie', 'biopsie hépatique']
    },
    {
        'Maladie': 'Tumeur stromale gastro-intestinale',
        'Symptomes': ['Douleur abdominale ou inconfort',
                        'Saignements gastro-intestinaux',
                        'Anémie',
                        'Fatigue',
                        'Sensation de plénitude ou de satiété précoce'],
        'Traitement': ['chirurgie', 'thérapie ciblée', 'chimiothérapie'],
        'Examens': ['endoscopie gastro-intestinale', 'biopsie', 'imagerie médicale']
    }, 
    {
        'Maladie': 'Lymphomes',
        'Symptomes': ["Gonflement indolore des ganglions lymphatiques (notamment au cou et aux aisselles ou à l'aine)",
                        'Fatigue',
                        'Fièvre inexpliquée',
                        'Perte de poids inexpliquée',
                        'Sudation nocturne excessive'],
        'Traitement': ['chimiothérapie', 'immunothérapie', 'radiothérapie'],
        'Examens': ["biopsie d'un ganglion lymphatique", 'scanner', 'IRM']
    },
    {
        'Maladie': 'Carcinome hépatocellulaire',
        'Symptomes': ['Douleur abdominale',
                        "Gonflement de l'abdomen",
                        'Perte de poids inexpliquée',
                        'Fatigue',
                        'Jaunissement de la peau et des yeux (ictère)'],
        'Traitement': ['chirurgie', 'radiofréquence', 'greffe de foie'],
        'Examens': ['échographie du foie', 'scanner', 'IRM']
    },
    {
        'Maladie': 'Sarcome de Kaposi',
        'Symptomes': ['Lésions cutanées violettes, rouges ou brunes',
                        'Gonflement des jambes ou des pieds',
                        'Douleurs dans les jambes',
                        'Essoufflement'],
        'Traitement': ['thérapie antirétrovirale', 'chirurgie', 'radiothérapie'],
        'Examens': ['biopsie cutanée', 'analyse sanguine', 'imagerie médicale']
    },
    {
        'Maladie': 'Cancer du larynx',
        'Symptomes': ['Enrouement persistant',
                        'Changement dans la voix',
                        "Douleur à la gorge ou à l'oreille",
                        'Difficulté à avaler'],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['laryngoscopie', 'biopsie', 'imagerie médicale']
    },
    {
        'Maladie': 'Mésothéliome',
        'Symptomes': ['Essoufflement',
                        'Douleur thoracique',
                        'Toux persistante',
                        'Perte de poids inexpliquée',
                        'Fatigue'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'thérapie ciblée'],
        'Examens': ['scanner thoracique', 'biopsie', 'imagerie médicale']
    },
    {
        'Maladie': "Cancer de l'œsophage",
        'Symptomes': ['Difficulté à avaler (dysphagie)',
                        'Douleur ou sensation de brûlure derrière le sternum',
                        'Perte de poids inexpliquée',
                        'Toux persistante'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['endoscopie digestive haute', 'biopsie', 'imagerie médicale']
    },
    {
        'Maladie': 'Ostéosarcome',
        'Symptomes': ['Douleur osseuse, généralement dans les bras ou les jambes',
                        'Gonflement autour des os affectés',
                        'Fractures osseuses fréquentes',
                        'Masse palpable sur les os', "fractures spontanées"],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['radiographie', 'IRM', 'biopsie']
    },
    {
        'Maladie': "Cancer de l'ovaire",
        'Symptomes': ['Gonflement abdominal',
                        'Douleur pelvienne ou abdominale',
                        "Besoin fréquent d'uriner",
                        'Sensation de satiété rapide en mangeant'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'hormonothérapie'],
        'Examens': ['échographie pelvienne', 'tomographie par émission de positons (TEP)', 'biopsie']
    },
    {
        'Maladie': 'Cancer du pancréas',
        'Symptomes': ['Douleur abdominale ou lombaire',
                        'Perte de poids inexpliquée',
                        'Jaunissement de la peau et des yeux (ictère)',
                        "Perte d'appétit"],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['scanner abdominal', 'IRM', 'biopsie']
    },
    {
        'Maladie': 'Cancer de la peau',
        'Symptomes': ['Changements dans la taille, la forme ou la couleur des grains de beauté ou des lésions cutanées',
                        'Saignements ou démangeaisons persistantes',
                        'Douleur',
                        'Croissance rapide'],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['biopsie de la peau', 'dermatoscopie', 'imagerie médicale']
    },
    {
        'Maladie': 'Cancer de la bouche',
        'Symptomes': ['Lésions buccales persistantes',
                        'Difficulté à mâcher ou à avaler',
                        "Douleur à l'oreille",
                        'Masse ou enflure dans la bouche ou le cou'],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['examen buccal', 'biopsie', 'imagerie médicale']
    },
     {
        'Maladie': 'Carcinome pulmonaire à petites cellules (cancer du poumon)',
        'Symptomes': ['Toux persistante',
                        'Essoufflement',
                        'Douleur thoracique',
                        'Perte de poids inexpliquée',
                        'Toux avec du sang'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['scanner thoracique', 'biopsie pulmonaire', 'tomographie par émission de positons (TEP)']
    },
    {
        'Maladie': 'Cancer de la prostate',
        'Symptomes': [ "Besoin fréquent d'uriner",
                        'Douleur ou sensation de brûlure pendant la miction',
                        'Flux urinaire faible ou interrompu',
                        "Sang dans l'urine ou le sperme",
                        'Douleur dans le dos, les hanches ou le bassin'],
        'Traitement': ['chirurgie', 'radiothérapie', 'hormonothérapie'],
        'Examens': ["dosage de l'antigène prostatique spécifique (PSA)", 'biopsie de la prostate', 'IRM pelvienne']
    },
    {
        'Maladie': 'Cancer du rein',
        'Symptomes': ['Douleur dans le dos ou sur le côté',
                        'Masse abdominale palpable',
                        "Sang dans l'urine",
                        'Fatigue',
                        'Perte de poids inexpliquée'],
        'Traitement': ['chirurgie', 'immunothérapie', 'thérapie ciblée'],
        'Examens': ['imagerie rénale (IRM et scanner)', 'urographie intraveineuse', 'biopsie rénale']
    },
    {
        'Maladie': 'Cancer du testicule',
        'Symptomes': ['Gonflement ou masse dans le testicule',
                        'Douleur ou inconfort dans le testicule ou le scrotum',
                        'Sensation de lourdeur dans le scrotum',
                        "Douleur ou inconfort dans le bas du dos ou l'aine"],
        'Traitement': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'Examens': ['examen physique', 'échographie testiculaire', 'biopsie']
    },
    {
        'Maladie': 'Cancer de la thyroïde',
        'Symptomes': [ 'Gonflement ou bosse dans le cou',
                            'Changements dans la voix',
                            'Difficulté à avaler',
                            'Douleur dans le cou'],
        'Traitement': ['chirurgie', 'thyroïdectomie', 'radiothérapie'],
        'Examens': ['échographie thyroïdienne', 'scintigraphie thyroïdienne', "biopsie à l'aiguille fine"]
    },
    {
        'Maladie': 'Sarcome des tissus mous',
        'Symptomes': ['Masse ou gonflement indolore',
                        'Douleur persistante',
                        'Saignements inhabituels',
                        'Fatigue'],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['biopsie', 'IRM', 'scanner']
    },
    {
        'Maladie': 'Carcinome de la vessie',
        'Symptomes': ["Sang dans l'urine",
                    'Douleur à la miction',
                    "Besoin fréquent d'uriner",
                    'Douleur pelvienne'],
        'Traitement': ['chirurgie', 'chimiothérapie', 'immunothérapie'],
        'Examens': ['cystoscopie', 'urographie intraveineuse', 'biopsie de la vessie']
    },
    {
        'Maladie': 'Myélome',
        'Symptomes': ['Douleurs osseuses, surtout dans le dos ou les côtes',
                        'Fatigue',
                        'Fractures osseuses fréquentes',
                        'Infections récurrentes',
                        'Perte de poids inexpliquée'],
        'Traitement': ['chimiothérapie', 'transplantation de cellules souches', 'thérapie ciblée'],
        'Examens': ['électrophorèse des protéines sériques et urinaires', 'biopsie de moelle osseuse', 'imagerie médicale']
    },
    {
        'Maladie': 'Plasmocytome',
        'Symptomes': ['Douleurs osseuses localisées',
                        'Masse palpable sur les os',
                        'Fractures osseuses',
                        'Fatigue'],
        'Traitement': ['chirurgie', 'radiothérapie', 'chimiothérapie'],
        'Examens': ['biopsie', 'imagerie médicale', 'analyse sanguine']
    }
]
liste =[]
for data in data_d:
        liste.append(data['Maladie'])
len(liste)

data_l = [{
        'nom': 'Pneumonie',
        'symptomes': ['fièvre', "toux avec expectorations", "douleur thoracique", "difficulté à respirer"],
        'traitements': ['antibiotiques','antipyrétiques', 'repos'],
        'examens': ['radiographie pulmonaire', 'analyse de sang']
    },
    {
        'nom': 'Bronchite',
        'symptomes': ['toux persistante', 'production de mucosités', 'fatigue fièvre légère'],
        'traitements': ['repos', 'antitussifs', 'hydratation'],
        'examens': ['analyse de crachat', 'radiographie pulmonaire si nécessaire']
    },
    {
        'nom': 'Otite',
        'symptomes': ["douleur à l'oreille", "fièvre perte auditive temporaire"],
        'traitements': ['antibiotiques (si bactérienne)', "analgésiques", "gouttes auriculaires"],
        'examens': ["examen de l'oreille par un professionnel de la santé"]
    },
    {
        'nom': 'Méningite',
        'symptomes': ['raideur de la nuque', 'maux de tête sévères fièvre', 'confusion'],
        'traitements': ['antibiotiques', 'traitement de support', 'hospitalisation'],
        'examens': ['ponction lombaire','analyse de liquide céphalorachidien']
    },
    {
        'nom': 'Angine',
        'symptomes': ['maux de gorge', 'difficulté à avaler', 'fièvre ganglions enflés'],
        'traitements': ['antibiotiques (si bactérienne)', 'antalgiques', 'gargarismes'],
        'examens': ['examen de la gorge', 'analyse de la gorge pour détecter une infection bactérienne']
    },
    {
        'nom': 'Infections urinaires',
        'symptomes': ['brûlures en urinant', "besoin fréquent d'uriner", 'douleur pelvienne'],
        'traitements': ['antibiotiques', 'hydratation', 'repos'],
        'examens': ["analyse d'urine", "culture d'uriner"]
    },
    {
        'nom': 'Panaris',
        'symptomes': ["rougeur et gonflement autour de l'ongle", "douleur pulsatile"],
        'traitements': ["drainage de l'abcès"," antibiotiques", 'repos'],
        'examens': ['examen clinique', 'culture du pus si nécessaire']
    },
    {
        'nom': 'Septicémie',
        'symptomes': ['fièvre élevée', 'confusion', 'frissons', 'hypotension'],
        'traitements': ['antibiotiques par voie intraveineuse', 'soutien hémodynamique', 'soins intensifs'],
        'examens': ['hémoculture', 'analyse de sang approfondie']
    },
    {
        'nom': 'Varicelle',
        'symptomes': ['éruption cutanée avec des démangeaisons', 'fièvre', 'maux de tête', 'fatigue'],
        'traitements': ['antihistaminiques', 'antipyrétiques', "calamine pour les démangeaisons"],
        'examens': ['les symptômes sont souvent suffisants pour le diagnostic']
    },
    {
        'nom': 'Rougeole',
        'symptomes': ['éruption cutanée', 'fièvre élevée', 'toux', 'conjonctivite'],
        'traitements': ['repos', 'antipyrétiques', "vitamine A en cas de carence"],
        'examens': ['analyse de sang', "observation des symptômes caractéristiques"]
    },
    {
        'nom': 'Rubéole',
        'symptomes': ['éruption cutanée légère', 'fièvre légère', 'ganglions enflés'],
        'traitements': ['repos', "gestion des symptômes"],
        'examens': ['analyse de sang pour détecter les anticorps']
    },
    {
        'nom': 'Coqueluche',
        'symptomes': ['toux sévère par accès', 'éternuements', 'épuisement', 'vomissements après la toux'],
        'traitements': ['antibiotiques', "vaccin de rappel"],
        'examens': ['prélèvement de sécrétions nasales pour la culture']
    },
    {
        'nom': 'Oreillons',
        'symptomes': ['gonflement des glandes salivaires', 'fièvre', "douleur en mangeant ou en buvant"],
        'traitements': ['repos', "antipyrétiques", "compresses chaudes sur les glandes enflées"],
        'examens': ['observation des symptômes caractéristiques']
    },
    {
        'nom': 'Maladies cardiovasculaires',
        'symptomes': ['douleur thoracique', 'essoufflement', 'fatigue', 'palpitations'],
        'traitements': ['médicaments antiplaquettaires', 'médicaments pour abaisser la pression artérielle'],
        'examens': ['électrocardiogramme (ECG)', 'échocardiographie']
    },
    {
        'nom': 'Accidents vasculaires cérébraux',
        'symptomes': ['engourdissement', 'faiblesse soudaine', 'confusion', 'maux de tête sévères'],
        'traitements': ['thérapie de revascularisation', 'médicaments anticoagulants'],
        'examens': ['tomographie par ordinateur (TDM) du cerveau', 'imagerie par résonance magnétique (IRM)']
    },
    {
        'nom': 'Cancer',
        'symptomes': ['perte de poids inexpliquée', 'bosse anormale', 'fatigue persistante'],
        'traitements': ['chirurgie', 'chimiothérapie', 'radiothérapie'],
        'examens': ['biopsie', 'imagerie médicale']
    },
    {
        'nom': 'Diabète',
        'symptomes': ['soif excessive', 'mictions fréquentes', 'fatigue', 'vision floue'],
        'traitements': ['insuline', 'médicaments hypoglycémiants'],
        'examens': ['test de glycémie à jeun', "test d'hémoglobine A1c"]
    },
    {
        'nom': 'Arthrite',
        'symptomes': ['douleur articulaire', 'gonflement', 'raideur', 'fatigue'],
        'traitements': ['anti-inflammatoires', 'physiothérapie'],
        'examens': ['analyse de sang', 'imagerie des articulations']
    },
    {
        'nom': 'Maux de dos',
        'symptomes': ['douleur dans le bas du dos', 'raideur musculaire', 'douleur irradiant vers les jambes'],
        'traitements': ['repos', 'physiothérapie', 'médicaments analgésiques'],
        'examens': ['radiographie de la colonne vertébrale', 'IRM']
    },
    {
        'nom': 'Dépression chronique',
        'symptomes': ['tristesse persistante', "perte d'intérêt", 'fatigue', 'troubles du sommeil'],
        'traitements': ['psychothérapie', 'antidépresseurs'],
        'examens': ['entretien avec un professionnel de la santé mentale', 'questionnaires de dépistage']
    },
        {
        'nom': 'Paludisme',
        'symptomes': ['fièvre', 'frissons', 'maux de tête', 'fatigue'],
        'traitements': ['antipaludiques', 'repos'],
        'examens': ['test de dépistage du paludisme']
    },
    {
        'nom': 'Bilharziose (Schistosomiase)',
        'symptomes': ['fièvre', 'toux', 'éruption cutanée', 'douleurs musculaires'],
        'traitements': ['antiparasitaires', 'antibiotiques'],
        'examens': ["examen microscopique des selles ou de l'urine"]
    },
    {
        'nom': 'Trypanosomiase africaine',
        'symptomes': ['fièvre', 'fatigue', 'maux de tête', 'ganglions enflés'],
        'traitements': ['médicaments antiparasitaires', 'antibiotiques'],
        'examens': ['test sanguin']
    },
    {
        'nom': 'Maladie de Chagas',
        'symptomes': ['fièvre', 'fatigue', 'gonflement des paupières', 'troubles cardiaques'],
        'traitements': ['médicaments antiparasitaires', 'médicaments pour le cœur'],
        'examens': ['test sanguin', 'électrocardiogramme (ECG)']
    },
    {
        'nom': 'Leishmaniose (Kala Azar)',
        'symptomes': ['fièvre prolongée', 'perte de poids', 'anémie', 'enflure de la rate et du foie'],
        'traitements': ['médicaments antiparasitaires', 'traitement de support'],
        'examens': ['aspiration de moelle osseuse']
    }
]

liste =[]
for data in data_l:
        liste.append(data['nom'])
len(liste)

### Creation du dataset  

dataset = []

for data1 in data_d:
        dataset.append({
            'Maladies': data1['Maladie'],
            'Symptomes': ','.join([str(i) for i in data1["Symptomes"]]),
            'Traitement': ",".join(data1['Traitement']),
            'Examens': ",".join(data1['Examens'])
        })
         
for data in data_l:
        dataset.append({
            'Maladies': data['nom'],
            'Symptomes': ",".join(data["symptomes"]),
            'Traitement': ",".join(data['traitements']),
            'Examens': ",".join(data['examens'])
        }) 
        
df = pd.DataFrame(dataset)
df.head()

### Exportation

# Exportation du DataFrame en fichier CSV et excel

df.to_excel('Ressources/RessourcesCollecte2/Dataset_Mladies.xlsx', index=False)

df.to_csv('Ressources/RessourcesCollecte2/Dataset_Maladies.csv', index=False)

print("DataFrame exporté avec succès")
