==> MEDICAL AI ANALYSE/3 - ANALYSE DE DONNEES ET DATA CLEANING.ipynb <==
# <center>ANALYSE DE DONNEES ET DATA CLEANING (nettoyage de données)</center> 

### OBJECTIF: 
        - Améliorer la Compréhension des Données
        - Gestion des Valeurs Manquantes
        - Identification et Gestion des Doublons
        - Préparation pour l'Analyse et la visualisation de donnees    
---

### Importations

import pandas as pd

### Lecture du dataset

df = pd.read_csv("Ressources/RessourcesCollecte2/Dataset_Maladies.csv")
df

### Verifions le type de données et le nombre de valeurs manquantes

df.info()

valeurs_manquantes = df.isnull().sum()
valeurs_manquantes

df.dtypes

### CONCLUSION 1: nous n'avons pas de valeurs manquantes et le type de nos données est cohérent

### Verifions le nombre de valeurs unique sur chaque colonne

len(df) - df.nunique()

### Verifions le nombre de doublons dans notre jeux de donnees

df.duplicated().sum()

### CONCLUSION 2: notre dataset ne possede pas de doublons

### Nettoyons les valeurs contenues dans les colonnes Symptomes,Traitement et examens

def netoyage(dfr):
    dfr = dfr.str.lower()
    dfr = dfr.str.strip()
    return dfr

df = df.apply(lambda x: netoyage(x))

df.head()

### Exportation

# Exportation du DataFrame en fichier CSV, json et excel

df.to_excel('Ressources/RessourcesAnalyse/DatasetClean.xlsx', index=False)

df.to_csv('Ressources/RessourcesAnalyse/DatasetClean.csv', index=False)

df.to_json('Ressources/RessourcesAnalyse/DatasetClean.json', orient='records', lines=True)

print("DataFrame exporté avec succès.")
