==> MEDICAL AI ANALYSE/5 - NLP SPACY.ipynb <==
### Importations

import pandas as pd
import numpy as np
import re
from collections import OrderedDict
from nltk.tokenize import word_tokenize
from unidecode import unidecode
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer,PorterStemmer

### Lecture du dataset

df = pd.read_csv("Ressources/RessourcesAnalyse/DatasetClean.csv")
df.head()

## I - Nettoyage de données 

stop_words = stopwords.words('french')
lemmatizer = WordNetLemmatizer()

ps = PorterStemmer()

new_stopwords =["facilement","tres","pendant"]

stop_words.extend(new_stopwords)

def Cleaning(text):
    
    def remove_s(word):
        if word.endswith('s'):
            return word[:-1]
        if word.endswith('ment'):
            return word[:-4]
        else:
            return word
        
    liste2 = ["abdomin ","abdomen","abdominal"]
    liste1 = ["douleur","mal"]
        
    # Supprimer des ponctuations
    text = re.sub("[^\w\s]", " ", text)
   
    text = ' '.join([lemmatizer.lemmatize(unidecode(word.lower())) for word in word_tokenize(text) if word.isalnum() and word.lower() not in stop_words])
    
    # supprimer les chiffres
    text = re.sub("\d", "", text)
    
    #text = [remove_s(word) for word in text]
    
    text = text.replace("ee ","e ")
    text = text.replace("cutane","peau")
    
    for i in liste2:
        text = text.replace(i,"ventre ")
        
    text = list(OrderedDict.fromkeys(text.split()))
    
    text = ' '.join([remove_s(word) for word in text])
    
    for i in liste1:
        text = text.replace(i,"maux")
    
    text = text.replace(" e "," ")
    return text



df["Symptomes_lem_token"] = df["Symptomes"].apply(lambda x: Cleaning(x))
df["Symptomes_lem_token"]

## III -  Calcul de similarité

liste_symptomes ="j'ai la fievres douala poid peau perte fatigue//. frissons yaya essouflement yaounde maux facilement , la ,   tête"
liste_symptomes = Cleaning(liste_symptomes)


import spacy

# Charger le modèle linguistique français avec des vecteurs de mots
nlp = spacy.load('fr_core_news_md')

similarite = []

for i in range(len(df["Symptomes_lem_token"])):
    
    liste_symptome = " ".join([i for i in df['Symptomes_lem_token'][i].split()  if i in liste_symptomes.split()])
    
    val = nlp(liste_symptome)
    
    doc = nlp(df['Symptomes_lem_token'][i])
    
    similarity = doc.similarity(val)
    
    similarite.append(similarity)

similarite = np.array(similarite)

df['Similarite_%'] = 100*similarite.flatten()

df['Similarite_%']

df_tri = df.sort_values(by='Similarite_%', ascending=False)

print(liste_symptomes)
df_tri
