==> MEDICAL AI ANALYSE/6 - Test.ipynb <==
## Importation

import pandas as pd
import numpy as np
import re
from nltk.tokenize import word_tokenize
from unidecode import unidecode
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer,PorterStemmer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
import joblib

df = pd.read_csv("Ressources/RessourcesAnalyse/DTfinal.csv")
df.head()

class NLTKSearch:
    def __init__(self,df):
        self.df = df
        self.tfidf_vectorizer = TfidfVectorizer()
        self.tfidf_matrix = self.tfidf_vectorizer.fit_transform(df["Symptomes_stem_token"]).toarray()
        self.stop_words = list(set(stopwords.words('french')))
        self.lemmatizer = WordNetLemmatizer()
        self.ps = PorterStemmer()
    
    def search(self, query):
        
        def Cleaning(text):
            
            new_stopwords =["facilement","tres","pendant"]

            self.stop_words.extend(new_stopwords)
    
            def remove_s(word):
                if word.endswith('s'):
                    return word[:-1]
                if word.endswith('ment'):
                    return word[:-4]
                else:
                    return word

            liste2 = ["abdomin ","abdomen","abdominal"]
            liste1 = ["douleur","mal"]

            # Supprimer des ponctuations
            text = re.sub("[^\w\s]", " ", text)
            
            text = ' '.join([self.ps.stem(unidecode(word.lower())) for word in word_tokenize(text) if word.isalnum() and word.lower() not in self.stop_words])

            # supprimer les chiffres
            text = re.sub("\d", "", text)

            #text = [remove_s(word) for word in text]

            text = text.replace("ee ","e ")
            text = text.replace("cutane","peau")

            for i in liste2:
                text = text.replace(i,"ventre ")

            text = list(set(text.split()))

            text = ' '.join([remove_s(word) for word in text])

            for i in liste1:
                text = text.replace(i,"maux")

            text = text.replace(" e "," ")
            return text
    
        processed_query = Cleaning(query)
        
        
        tif2 = TfidfVectorizer()

        cosine_similarite = []

        for i in range(len(df["Symptomes_stem_token"])):
            tfidf2 = tif2.fit_transform([df['Symptomes_stem_token'][i]]).toarray()
            cosine_similaritie = cosine_similarity(tfidf2, tif2.transform([processed_query]))[0]
            cosine_similarite.append(cosine_similaritie)
        
        cosine_similaritie = np.array(cosine_similarite)
        df['Similarites_%'] = 100*cosine_similaritie.flatten()
        df_tri = df.sort_values(by='Similarites_%', ascending=False)
        
        return df_tri

# Créer une instance du moteur de recherche
search_ = NLTKSearch(df)

# Sauvegarder le moteur de recherche avec joblib
joblib.dump(search_, 'Ressources/Joblib/Diagnostic_search.joblib')

# Charger le moteur de recherche depuis le fichier
Model = joblib.load('Ressources/Joblib/Diagnostic_search.joblib')

# Utiliser le moteur de recherche chargé
query_results = Model.search("j'ai la fievres douala poid peau perte fatigue//. frissons yaya essouflement yaounde maux facilement , la ,   tête")
query_results
